analysis/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
requests
google-genai
